Squashed commit of the following:
commit a15a4298dc5348376f501c6fb0c52aaa71b0670d
Author: SeanKim-416 <56223156+SeanKim-416@users.noreply.github.com>
Date:   Thu Nov 9 16:02:56 2023 +0900

    Refactor: createWarningText 함수 모듈화

commit 685dd76b45a5e99147f438f96a20aabf2c294adf
Author: SeanKim-416 <56223156+SeanKim-416@users.noreply.github.com>
Date:   Thu Nov 9 15:56:06 2023 +0900

    Refactor: 구조분해를 이용하여 코드 최적화, arrow function으로 변경

commit 816a06128141ef57d1768996e91f026d41473a34
Author: SeanKim-416 <56223156+SeanKim-416@users.noreply.github.com>
Date:   Thu Nov 9 15:53:29 2023 +0900

    Refactor: 재사용될 이벤트와 함수 모듈화

commit 6bfb40e6f574daf728945d94501dcd076ddf40b7
Author: SeanKim-416 <56223156+SeanKim-416@users.noreply.github.com>
Date:   Thu Nov 9 14:03:51 2023 +0900

    Refactor: focusin 이벤트 최적화
